+++ src/torch/cpu/ops_impl_cpu.h
@@ -28,7 +28,7 @@ void fill_impl(Tensor& a, Tensor values, int64_t dim);
 void permute_impl(Tensor& src, Tensor result, SizeType new_dims);
 
 void uniform_impl(Tensor& t, double mi, double ma);
-void uniform_int_impl(Tensor& t, int low, int high);
+void uniform_int_impl(Tensor& t, int64_t low, int64_t high);
 void abs_sum_impl(Tensor a, Tensor result);
 void sum_impl(Tensor a, Tensor result);
 void sum_impl(Tensor a, int64_t dim, Tensor result);
